fix: need cast to int

+++ src/pleocmd/pipe/val/Value.java
@@ -116,7 +116,8 @@ public abstract class Value {
 		if (c == StringValue.TYPE_CHAR) return StringValue.RECOMMENDED_TYPE;
 		if (c == BinaryValue.TYPE_CHAR) return BinaryValue.RECOMMENDED_TYPE;
 		throw new IOException(String.format(
-				"Invalid type identifier: 0x%02X at position %d", c, index));
+				"Invalid type identifier: 0x%02X at position %d", (int) c,
+				index));
 	}
 
 	public static int getAsciiTypeChar(final Value value) {
